requirements.txt
adbutils==1.0.2
apkutils2==1.0.0
Appium-Python-Client==1.0.2
atomicwrites==1.4.0
attrs==21.4.0
cffi==1.15.1
charset-normalizer==2.1.0
cigam==0.0.3
colorama==0.4.5
decorator==5.1.1
deprecation==2.1.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pycryptodome==3.15.0
pyelftools==0.28
pyparsing==3.0.9
pytest==7.1.2
pytest-testreport==1.1.5
requests==2.28.1
retry==0.9.2
selenium==3.141.0
tomli==2.0.1
urllib3==1.26.9
whichcraft==0.6.1
wincertstore==0.2
xmltodict==0.13.0
fp-browser-sdk==0.1